+++ ProgramFunctions/Fields.fnc.php
@@ -276,7 +276,8 @@ function GetFieldsForm( $table, $title, $RET, $extra_category_fields = [], $type
 			]
 		);
 
-		$delete_button = '<input type="button" value="' . AttrEscape( _( 'Delete' ) ) . '" onClick="ajaxLink(\'' . $delete_url . '\');" /> ';
+		$delete_button = '<input type="button" value="' . AttrEscape( _( 'Delete' ) ) .
+			'" onclick="' . AttrEscape( 'ajaxLink(' . json_encode( $delete_url ) . ');' ) . '" /> ';
 	}
 
 	ob_start();