+++ modules/Accounting/DailyTransactions.php
@@ -50,7 +50,7 @@ function _programMenu( $program )
 			'totals' => _( 'Daily Totals' ),
 		],
 		false,
-		'onchange="ajaxLink(\'' . $link . '\' + this.value);" autocomplete="off"',
+		'onchange="' . AttrEscape( 'ajaxLink(' . json_encode( $link ) . ' + this.value);' ) . '" autocomplete="off"',
 		false
 	);