+++ modules/Attendance/Percent.php
@@ -20,7 +20,7 @@ $report_select = SelectInput(
 		'true' => _( 'Average Attendance by Day' ),
 	],
 	false,
-	'onchange="ajaxLink(\'' . $report_link . '\' + this.value);" autocomplete="off"',
+	'onchange="' . AttrEscape( 'ajaxLink(' . json_encode( $report_link ) . ' + this.value);' ) . '" autocomplete="off"',
 	false
 );