+++ modules/Scheduling/ScheduleReport.php
@@ -29,7 +29,7 @@ $report_select = SelectInput(
 		'master' => _( 'Master Schedule Report' ),
 	],
 	false,
-	'onchange="ajaxLink(\'' . $report_link . '\' + this.value);" autocomplete="off"',
+	'onchange="' . AttrEscape( 'ajaxLink(' . json_encode( $report_link ) . ' + this.value);' ) . '" autocomplete="off"',
 	false
 );