+++ modules/School_Setup/MarkingPeriods.php
@@ -443,9 +443,10 @@ if ( ! $_REQUEST['modfunc'] )
 
 			if ( $can_delete )
 			{
-				$delete_URL = "'" . URLEscape( $mp_href . "&modfunc=delete" ) . "'";
+				$delete_URL = URLEscape( $mp_href . "&modfunc=delete" );
 
-				$delete_button = '<input type="button" value="' . AttrEscape( _( 'Delete' ) ) . '" onClick="javascript:ajaxLink(' . $delete_URL . ');" />';
+				$delete_button = '<input type="button" value="' . AttrEscape( _( 'Delete' ) ) .
+					'" onclick="' . AttrEscape( 'ajaxLink(' . json_encode( $delete_URL ) . ');' ) . '" />';
 			}
 		}
 	}