+++ modules/Students/Student.php
@@ -726,10 +726,11 @@ if (  ( UserStudentID()
 
 				if ( $can_delete )
 				{
-					$delete_URL = "'" . URLEscape( "Modules.php?modname=" . $_REQUEST['modname'] .
-						'&student_id=' . UserStudentID() . "&modfunc=delete" ) . "'";
+					$delete_URL = URLEscape( "Modules.php?modname=" . $_REQUEST['modname'] .
+						'&student_id=' . UserStudentID() . "&modfunc=delete" );
 
-					$delete_button = '<input type="button" value="' . AttrEscape( _( 'Delete' ) ) . '" onClick="javascript:ajaxLink(' . $delete_URL . ');" />';
+					$delete_button = '<input type="button" value="' . AttrEscape( _( 'Delete' ) ) .
+						'" onclick="' . AttrEscape( 'ajaxLink(' . json_encode( $delete_URL ) . ');' ) . '" />';
 				}
 			}
 		}