+++ modules/Users/includes/Schedule.inc.php
@@ -12,11 +12,13 @@ if ( GetTeacher( UserStaffID(), 'PROFILE', false ) === 'teacher' )
 		PopTable( 'footer' );
 	}
 
-	$all_schools_onclick_URL = "'" . ( $_REQUEST['all_schools'] == 'Y' ?
+	$all_schools_onclick_URL = ( $_REQUEST['all_schools'] == 'Y' ?
 		PreparePHP_SELF( $_REQUEST, [], [ 'all_schools' => '' ] ) :
-		PreparePHP_SELF( $_REQUEST, [], [ 'all_schools' => 'Y' ] ) ) . "'";
+		PreparePHP_SELF( $_REQUEST, [], [ 'all_schools' => 'Y' ] ) );
 
-	$input_all_schools = '<input type="checkbox" name="all_schools" value="Y" onclick="ajaxLink(' . $all_schools_onclick_URL . ');"' . ( $_REQUEST['all_schools'] == 'Y' ? 'checked' : '' ) . ' />';
+	$input_all_schools = '<input type="checkbox" name="all_schools" value="Y" onclick="' .
+		AttrEscape( 'ajaxLink(' . json_encode( $all_schools_onclick_URL ) . ');' ) . '"' .
+		( $_REQUEST['all_schools'] == 'Y' ? 'checked' : '' ) . ' />';
 
 	DrawHeader( '<label>' . $input_all_schools . ' ' . _( 'List Courses For All Schools' ) . '</label>' );